chore: move type enums to their own module
